src/requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.5
behave==1.2.6
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.3.0
execnet==2.1.1
Faker==37.8.0
fasteners==0.20
filelock==3.19.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mycdp==1.2.0
outcome==1.3.0.post0
packaging==25.0
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.6
pdbp==1.7.1
platformdirs==4.4.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pynose==1.5.5
pyotp==2.9.0
pypiwin32==223
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.4.2
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==16.0.1
pytest-seleniumbase==4.41.8
pytest-xdist==3.8.0
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
sbvirtualdisplay==1.4.0
selenium==4.35.0
seleniumbase==4.41.8
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tabcompleter==1.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
winsdk==1.0.0b10
wsproto==1.2.0